move recipients page to separate component
