+++ src/components/MultiSig/components/CSVUpload/types.ts
@@ -1,5 +1,8 @@
+import {Account} from '@cosmjs/stargate';
 export type RecipientAmount = { recipient: string; amount: number; denom: string; };
 
 export type CSVUploadProps = {
 	onParseData: (data: RecipientAmount[]) => void;
 };
+
+export type AddRecipientPageProps = {multiSigAccount: Account, setFinalizedRecipients: (recipientAmounts: RecipientAmount[]) => void};
